requirements.txt
# Core - IB API
ib_async>=0.9.86

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Config & Utils
python-dotenv>=1.0.0
loguru>=0.7.0
httpx>=0.27.0

# API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
